Move duplicated code from e2e tests to src folder
